backend-job-matcher/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.21.0
torch>=2.0.0
pymongo>=4.6.0
PyPDF2>=3.0.0
python-docx>=1.1.0


